Components renaming and switching to @apollo/client gql imports
