+++ client/src/components/settings/settingsEmails.vue
@@ -51,7 +51,7 @@
 </template>
 
 <script>
-import gql from 'graphql-tag'
+import { gql } from '@apollo/client'
 import { get } from 'vuex-pathify'
 import { isEmail } from 'validator'
